\name{wahc-package}
\alias{wahc-package}
\alias{wahc}
\docType{package}
\title{
Autocorrelation and Heteroskedasticity Correction in Fixed Effect Panel Data Model
}
\description{
Fit the fixed effect panel data model with heteroskedasticity and autocorrelation correction.
}
\details{
\tabular{ll}{
Package: \tab wahc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-23\cr
License: \tab GPL-3\cr
}
In this package, we apply the consistent (HAC) standard errors to deal with both problems heteroskedasticity and autocorrelation in the fixed effect panel data regression.
}
\author{
Zaghdoudi Taha

Zaghdoudi Taha  <zedtaha@gmail.com>
}
\references{

Bhargava A, Franzini L, Narendranathan W (1982). Serial Correlation and the Fixed Effects
Model.\emph{Review of Economic Studies},\bold{49}, pp.533--554.

Drukker D (2003), Testing for Serial Correlation in Linear Panel-Data Models.\emph{The Stata Journal},\bold{3}, pp. 168--177.

MacKinnon J, White H (1985), Some Heteroskedasticity-Consistent Covariance Matrix Es-
timators With Improved Finite Sample Propertie,\emph{Journal of econometrics}, \bold{29}, pp.305--325.

Zeileis A (2004). Econometric Computing With HC and HAC Covariance Matrix Estimators.\emph{Journal of Statistical Software}, \bold{11},pp. 1--17.


}

